+++ next.config.ts
@@ -5,7 +5,6 @@ import type { NextConfig } from "next";
  */
 const nextConfig: NextConfig = {
   /* config options here */
-  output: "export",
   images: {
     unoptimized: true,
   },